chore: 자잘한 오류 고침
