+++ client/src/App.js
@@ -1,12 +1,13 @@
 import './App.css';
 import Header from './components/header';
 import NavBar from './components/NavBar';
-
+import LogIn from './pages/LoginPage';
 function App() {
   return (
     <div className="App">
       <Header />
       <NavBar />
+      <LogIn />
       test
     </div>
   );
