+++ client/src/pages/LoginPage.jsx
@@ -1,6 +1,5 @@
 import '../styles/Login.css';
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import LoginFunc from '../auth/LoginFunc';
 
 const LogIn = () => {
@@ -24,7 +23,7 @@ const LogIn = () => {
         try {
             const result = await LoginFunc(id, password);
             if (result) {
-                window.location.href = '메인 주소';
+                window.location.href = '/';
             }
         }
         catch (err) {
@@ -32,33 +31,33 @@ const LogIn = () => {
         }
     }
 
-    const navigate = useNavigate();
-
     return (
         <div className='page_container'>  
-            <div className='logo_container'>
-                <img className="logo_leaf" src={require('../assets/logo_only_image.png')} alt='logo'/>
-            </div>
-            <div className='form_container'>
-                <form className='login_form' onSubmit={submitHandler}>
-                    <div className='id'>
-                        <label htmlFor="id">ID</label>
-                        <input id="id" onChange={emailHandler}></input>
-                    </div>
-                    <div className='password'>
-                        <label htmlFor="password">Password</label>
-                        <input id="password" onChange={passwordHandler} />
-                    </div>
-                    <div className='submit_button' onclick={navigate('/')}>
-                        <input className='submit' type="submit" value="로그인" />
-                    </div>
-                </form>
-            </div>
-            <div className='links'>
+            <div className='container'>  
+                <div className='logo_container'>
+                    <img className="logo_leaf" src={require('../assets/logo_only_image.png')} alt='logo'/>
+                </div>
+                <div className='form_container'>
+                    <form className='login_form'>
+                        <div className='id'>
+                            <label htmlFor="id">ID</label>
+                            <input id="id" onChange={emailHandler}></input>
+                        </div>
+                        <div className='password'>
+                            <label htmlFor="password">Password</label>
+                            <input id="password" onChange={passwordHandler} />
+                        </div>
+                        <div className='submit_button' onSubmit={submitHandler}>
+                            <input className='submit' type="submit" value="로그인" />
+                        </div>
+                    </form>
+                </div>
+                <div className='links'>
                     <a>비밀번호 찾기 </a>
                     <div className='center'> | </div>
                     <a>회원가입</a>
                 </div>
+            </div>
         </div>
     )
 }
